Shows how to access the server
This golang file will successfully access the server and get a packet.  It's pretty easy to modify this to get multiple packets, but this file just serves as a proof of concept for now.  I have a good idea how to decipher the packets as well.  This was made easy thanks to: http://atcswiki.greatlakesnetworking.net/bin/view/Main/NetworkConsiderations